+++ packages/remix-dev/config.ts
@@ -4,16 +4,17 @@ import { pathToFileURL } from "node:url";
 import fse from "fs-extra";
 import getPort from "get-port";
 import NPMCliPackageJson from "@npmcli/package-json";
+import { coerce } from "semver";
 
 import type { RouteManifest, DefineRoutesFunction } from "./config/routes";
 import { defineRoutes } from "./config/routes";
 import { defineConventionalRoutes } from "./config/routesConvention";
 import { ServerMode, isValidServerMode } from "./config/serverModes";
-import { serverBuildVirtualModule } from "./compiler/virtualModules";
-import { writeConfigDefaults } from "./compiler/utils/tsconfig/write-config-defaults";
+import { writeConfigDefaults } from "./config/write-tsconfig-defaults";
+import { serverBuildVirtualModule } from "./compiler/serverjs/virtualModules";
 import { flatRoutes } from "./config/flat-routes";
 import { getPreferredPackageManager } from "./cli/getPreferredPackageManager";
-import { warnOnce } from "./compiler/warnings";
+import { warnOnce } from "./warnOnce";
 
 export interface RemixMdxConfig {
   rehypePlugins?: any[];
@@ -56,6 +57,7 @@ interface FutureConfig {
   unstable_vanillaExtract: boolean | VanillaExtractOptions;
   v2_errorBoundary: boolean;
   v2_meta: boolean;
+  v2_normalizeFormMethod: boolean;
   v2_routeConvention: boolean;
 }
 
@@ -401,12 +403,16 @@ export async function readConfig(
     }
   }
 
+  if (appConfig.serverBuildTarget) {
+    warnOnce(serverBuildTargetWarning, "v2_serverBuildTarget");
+  }
+
   if (!appConfig.future?.v2_errorBoundary) {
     warnOnce(errorBoundaryWarning, "v2_errorBoundary");
   }
 
-  if (appConfig.serverBuildTarget) {
-    warnOnce(serverBuildTargetWarning, "v2_serverBuildTarget");
+  if (!appConfig.future?.v2_normalizeFormMethod) {
+    warnOnce(formMethodWarning, "v2_normalizeFormMethod");
   }
 
   let isCloudflareRuntime = ["cloudflare-pages", "cloudflare-workers"].includes(
@@ -471,7 +477,49 @@ export async function readConfig(
   if (userEntryServerFile) {
     entryServerFile = userEntryServerFile;
   } else {
-    if (!deps["isbot"]) {
+    let serverRuntime = deps["@remix-run/deno"]
+      ? "deno"
+      : deps["@remix-run/cloudflare"]
+      ? "cloudflare"
+      : deps["@remix-run/node"]
+      ? "node"
+      : undefined;
+
+    if (!serverRuntime) {
+      let serverRuntimes = [
+        "@remix-run/deno",
+        "@remix-run/cloudflare",
+        "@remix-run/node",
+      ];
+      let formattedList = disjunctionListFormat.format(serverRuntimes);
+      throw new Error(
+        `Could not determine server runtime. Please install one of the following: ${formattedList}`
+      );
+    }
+
+    let clientRenderer = deps["@remix-run/react"] ? "react" : undefined;
+
+    if (!clientRenderer) {
+      throw new Error(
+        `Could not determine renderer. Please install the following: @remix-run/react`
+      );
+    }
+
+    let maybeReactVersion = coerce(deps.react);
+    if (!maybeReactVersion) {
+      let react = ["react", "react-dom"];
+      let list = conjunctionListFormat.format(react);
+      throw new Error(
+        `Could not determine React version. Please install the following packages: ${list}`
+      );
+    }
+
+    let type: "stream" | "string" =
+      maybeReactVersion.major >= 18 || maybeReactVersion.raw === "0.0.0"
+        ? "stream"
+        : "string";
+
+    if (!deps["isbot"] && type === "stream") {
       console.log(
         "adding `isbot` to your package.json, you should commit this change"
       );
@@ -493,41 +541,35 @@ export async function readConfig(
       });
     }
 
-    let serverRuntime = deps["@remix-run/deno"]
-      ? "deno"
-      : deps["@remix-run/cloudflare"]
-      ? "cloudflare"
-      : deps["@remix-run/node"]
-      ? "node"
-      : undefined;
-
-    if (!serverRuntime) {
-      let serverRuntimes = [
-        "@remix-run/deno",
-        "@remix-run/cloudflare",
-        "@remix-run/node",
-      ];
-      let formattedList = listFormat.format(serverRuntimes);
-      throw new Error(
-        `Could not determine server runtime. Please install one of the following: ${formattedList}`
-      );
-    }
-
-    entryServerFile = `entry.server.${serverRuntime}.tsx`;
+    entryServerFile = `${serverRuntime}/entry.server.${clientRenderer}-${type}.tsx`;
   }
 
   if (userEntryClientFile) {
     entryClientFile = userEntryClientFile;
   } else {
-    let clientRuntime = deps["@remix-run/react"] ? "react" : undefined;
+    let clientRenderer = deps["@remix-run/react"] ? "react" : undefined;
 
-    if (!clientRuntime) {
+    if (!clientRenderer) {
       throw new Error(
         `Could not determine runtime. Please install the following: @remix-run/react`
       );
     }
 
-    entryClientFile = `entry.client.${clientRuntime}.tsx`;
+    let maybeReactVersion = coerce(deps.react);
+    if (!maybeReactVersion) {
+      let react = ["react", "react-dom"];
+      let list = conjunctionListFormat.format(react);
+      throw new Error(
+        `Could not determine React version. Please install the following packages: ${list}`
+      );
+    }
+
+    let type: "stream" | "string" =
+      maybeReactVersion.major >= 18 || maybeReactVersion.raw === "0.0.0"
+        ? "stream"
+        : "string";
+
+    entryClientFile = `entry.client.${clientRenderer}-${type}.tsx`;
   }
 
   let entryClientFilePath = userEntryClientFile
@@ -537,6 +579,10 @@ export async function readConfig(
   let entryServerFilePath = userEntryServerFile
     ? path.resolve(appDirectory, userEntryServerFile)
     : path.resolve(defaultsDirectory, entryServerFile);
+
+  if (appConfig.browserBuildDirectory) {
+    warnOnce(browserBuildDirectoryWarning, "browserBuildDirectory");
+  }
 
   let assetsBuildDirectory =
     appConfig.assetsBuildDirectory ||
@@ -633,6 +679,7 @@ export async function readConfig(
     unstable_vanillaExtract: appConfig.future?.unstable_vanillaExtract ?? false,
     v2_errorBoundary: appConfig.future?.v2_errorBoundary === true,
     v2_meta: appConfig.future?.v2_meta === true,
+    v2_normalizeFormMethod: appConfig.future?.v2_normalizeFormMethod === true,
     v2_routeConvention: appConfig.future?.v2_routeConvention === true,
   };
 
@@ -722,6 +769,8 @@ const resolveServerBuildPath = (
 
   // retain deprecated behavior for now
   if (appConfig.serverBuildDirectory) {
+    warnOnce(serverBuildDirectoryWarning, "serverBuildDirectory");
+
     serverBuildPath = path.join(appConfig.serverBuildDirectory, "index.js");
   }
 
@@ -732,16 +781,63 @@ const resolveServerBuildPath = (
   return path.resolve(rootDirectory, serverBuildPath);
 };
 
-// @ts-expect-error available in node 12+
-// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/ListFormat#browser_compatibility
-let listFormat = new Intl.ListFormat("en", {
+// adds types for `Intl.ListFormat` to the global namespace
+// we could also update our `tsconfig.json` to include `lib: ["es2021"]`
+declare namespace Intl {
+  type ListType = "conjunction" | "disjunction";
+
+  interface ListFormatOptions {
+    localeMatcher?: "lookup" | "best fit";
+    type?: ListType;
+    style?: "long" | "short" | "narrow";
+  }
+
+  interface ListFormatPart {
+    type: "element" | "literal";
+    value: string;
+  }
+
+  class ListFormat {
+    constructor(locales?: string | string[], options?: ListFormatOptions);
+    format(values: any[]): string;
+    formatToParts(values: any[]): ListFormatPart[];
+    supportedLocalesOf(
+      locales: string | string[],
+      options?: ListFormatOptions
+    ): string[];
+  }
+}
+
+let conjunctionListFormat = new Intl.ListFormat("en", {
   style: "long",
   type: "conjunction",
 });
 
-export let serverBuildTargetWarning = `⚠️ DEPRECATED: The "serverBuildTarget" config option is deprecated. Use a combination of "publicPath", "serverBuildPath", "serverConditions", "serverDependenciesToBundle", "serverMainFields", "serverMinify", "serverModuleFormat" and/or "serverPlatform" instead.`;
+let disjunctionListFormat = new Intl.ListFormat("en", {
+  style: "long",
+  type: "disjunction",
+});
 
-export let flatRoutesWarning = `⚠️ DEPRECATED: The old nested folders route convention has been deprecated in favor of "flat routes".  Please enable the new routing convention via the \`future.v2_routeConvention\` flag in your \`remix.config.js\` file.  For more information, please see https://remix.run/docs/en/main/file-conventions/route-files-v2.`;
+export let browserBuildDirectoryWarning =
+  "⚠️ DEPRECATED: The `browserBuildDirectory` config option is deprecated. " +
+  "Use `assetsBuildDirectory` instead.";
+
+export let serverBuildDirectoryWarning =
+  "⚠️ DEPRECATED: The `serverBuildDirectory` config option is deprecated. " +
+  "Use `serverBuildPath` instead.";
+
+export let serverBuildTargetWarning =
+  "⚠️ DEPRECATED: The `serverBuildTarget` config option is deprecated. Use a " +
+  "combination of `publicPath`, `serverBuildPath`, `serverConditions`, " +
+  "`serverDependenciesToBundle`, `serverMainFields`, `serverMinify`, " +
+  "`serverModuleFormat` and/or `serverPlatform` instead.";
+
+export let flatRoutesWarning =
+  "⚠️ DEPRECATED: The old nested folders route convention has been " +
+  "deprecated in favor of 'flat routes'.  Please enable the new routing " +
+  "convention via the `future.v2_routeConvention` flag in your " +
+  "`remix.config.js` file.  For more information, please see " +
+  "https://remix.run/docs/en/main/file-conventions/route-files-v2.";
 
 export const errorBoundaryWarning =
   "⚠️ DEPRECATED: The separation of `CatchBoundary` and `ErrorBoundary` has " +
@@ -750,3 +846,9 @@ export const errorBoundaryWarning =
   "behavior in Remix v1 via the `future.v2_errorBoundary` flag in your " +
   "`remix.config.js` file. For more information, see " +
   "https://remix.run/docs/route/error-boundary-v2";
+
+export const formMethodWarning =
+  "⚠️  DEPRECATED: Please enable the `future.v2_normalizeFormMethod` flag to " +
+  "prepare for the Remix v2 release. Lowercase `useNavigation().formMethod`" +
+  "values are being normalized to uppercase in v2 to align with the `fetch()` " +
+  "behavior.  For more information, see https://remix.run/docs/hooks/use-navigation";