+++ packages/remix-react/browser.tsx
@@ -14,7 +14,7 @@ import {
 import { deserializeErrors } from "./errors";
 import type { RouteModules } from "./routeModules";
 import { createClientRoutes } from "./routes";
-import { warnOnce } from "./warnings";
+import { logDeprecationOnce } from "./warnings";
 
 /* eslint-disable prefer-let/prefer-let */
 declare global {
@@ -140,14 +140,22 @@ if (import.meta && import.meta.hot) {
 export function RemixBrowser(_props: RemixBrowserProps): ReactElement {
   if (!router) {
     if (!window.__remixContext.future.v2_errorBoundary) {
-      warnOnce(
-        false,
+      logDeprecationOnce(
         "⚠️  DEPRECATED: The separation of `CatchBoundary` and `ErrorBoundary` has " +
           "been deprecated and Remix v2 will use a singular `ErrorBoundary` for " +
           "all thrown values (`Response` and `Error`). Please migrate to the new " +
           "behavior in Remix v1 via the `future.v2_errorBoundary` flag in your " +
           "`remix.config.js` file. For more information, see " +
           "https://remix.run/docs/route/error-boundary-v2"
+      );
+    }
+
+    if (!window.__remixContext.future.v2_normalizeFormMethod) {
+      logDeprecationOnce(
+        "⚠️  DEPRECATED: Please enable the `future.v2_normalizeFormMethod` flag to " +
+          "prepare for the Remix v2 release. Lowercase `useNavigation().formMethod`" +
+          "values are being normalized to uppercase in v2 to align with the `fetch()` " +
+          "behavior.  For more information, see https://remix.run/docs/hooks/use-navigation"
       );
     }
 
@@ -165,7 +173,18 @@ export function RemixBrowser(_props: RemixBrowserProps): ReactElement {
       };
     }
 
-    router = createBrowserRouter(routes, { hydrationData });
+    router = createBrowserRouter(routes, {
+      hydrationData,
+      future: {
+        // Pass through the Remix future flag to avoid a v1 breaking change in
+        // useNavigation() - users can control the casing via the flag in v1.
+        // useFetcher still always uppercases in the back-compat layer in v1.
+        // In v2 we can just always pass true here and remove the back-compat
+        // layer
+        v7_normalizeFormMethod:
+          window.__remixContext.future.v2_normalizeFormMethod,
+      },
+    });
   }
 
   // We need to include a wrapper RemixErrorBoundary here in case the root error