+++ templates/deno/app/entry.server.tsx
@@ -1,7 +1,13 @@
+/**
+ * By default, Remix will handle generating the HTTP Response for you.
+ * You are free to delete this file if you'd like to, but if you ever want it revealed again, you can run `npx remix reveal` ✨
+ * For more information, see https://remix.run/docs/en/main/file-conventions/entry.server
+ */
+
 import type { EntryContext } from "@remix-run/deno";
 import { RemixServer } from "@remix-run/react";
-import { renderToReadableStream } from "react-dom/server";
 import isbot from "isbot";
+import { renderToReadableStream } from "react-dom/server";
 
 export default async function handleRequest(
   request: Request,
@@ -9,7 +15,7 @@ export default async function handleRequest(
   responseHeaders: Headers,
   remixContext: EntryContext,
 ) {
-  const stream = await renderToReadableStream(
+  const body = await renderToReadableStream(
     <RemixServer context={remixContext} url={request.url} />,
     {
       signal: request.signal,
@@ -21,12 +27,11 @@ export default async function handleRequest(
   );
 
   if (isbot(request.headers.get("user-agent"))) {
-    await stream.allReady;
+    await body.allReady;
   }
 
   responseHeaders.set("Content-Type", "text/html");
-
-  return new Response(stream, {
+  return new Response(body, {
     headers: responseHeaders,
     status: responseStatusCode,
   });